+++ blocks/detail-viewer/src/components/renderers/FileRenderer/FileRenderer.jsx
@@ -5,18 +5,6 @@ import React from 'react';
 
 import styles from './FileRenderer.css';
 
-function getSrc(block, value) {
-  if (value instanceof Blob) {
-    return URL.createObjectURL(value);
-  }
-
-  if (block?.parameters?.fileBase) {
-    return `${new URL(`${block.parameters.fileBase}/${value}`, window.location.origin)}`;
-  }
-
-  return value;
-}
-
 /**
  * Render a string as is.
  */
@@ -33,6 +21,11 @@ export default class FileRenderer extends React.Component {
     block: PropTypes.shape().isRequired,
 
     /**
+     * The parameters passed in by the Appsemble block.
+     */
+    utils: PropTypes.shape().isRequired,
+
+    /**
      * The current value.
      */
     value: PropTypes.oneOfType([
@@ -46,8 +39,24 @@ export default class FileRenderer extends React.Component {
     value: null,
   };
 
+  getSrc = value => {
+    const { block, utils } = this.props;
+
+    if (value instanceof Blob) {
+      const url = URL.createObjectURL(value);
+      utils.addCleanup(() => URL.revokeObjectURL(url));
+      return url;
+    }
+
+    if (block?.parameters?.fileBase) {
+      return `${new URL(`${block.parameters.fileBase}/${value}`, window.location.origin)}`;
+    }
+
+    return value;
+  };
+
   render() {
-    const { block, field, value } = this.props;
+    const { field, value } = this.props;
 
     return (
       <React.Fragment>
@@ -64,7 +73,7 @@ export default class FileRenderer extends React.Component {
                   <img
                     alt={field.label || field.name}
                     className={styles.img}
-                    src={getSrc(block, field.repeatedName ? remapData(field.repeatedName, v) : v)}
+                    src={this.getSrc(field.repeatedName ? remapData(field.repeatedName, v) : v)}
                   />
                 </figure>
               );
@@ -72,11 +81,7 @@ export default class FileRenderer extends React.Component {
           </div>
         ) : (
           <figure className={classNames('image', styles.root)}>
-            <img
-              alt={field.label || field.name}
-              className={styles.img}
-              src={getSrc(block, value)}
-            />
+            <img alt={field.label || field.name} className={styles.img} src={this.getSrc(value)} />
           </figure>
         )}
       </React.Fragment>