+++ packages/server/api/components/schemas/index.js
@@ -6,6 +6,7 @@ import BlockVersion from './BlockVersion';
 import Color from './Color';
 import Error from './Error';
 import FlowPage from './FlowPage';
+import Health from './Health';
 import Organization from './Organization';
 import Page from './Page';
 import Resource from './Resource';
@@ -22,6 +23,7 @@ export default {
   Color,
   Error,
   FlowPage,
+  Health,
   Organization,
   Page,
   Resource,