+++ packages/server/index.js
@@ -2,6 +2,7 @@
 import { configureLogger, handleError } from '@appsemble/node-utils';
 import yargs from 'yargs';
 
+import * as health from './commands/health';
 import * as migrate from './commands/migrate';
 import * as start from './commands/start';
 
@@ -31,7 +32,13 @@ function main(argv) {
       describe: 'Decrease verbosity',
       type: 'count',
     })
+    .option('log-date', {
+      describe: 'Prepend the date to logging.',
+      type: 'boolean',
+      default: true,
+    })
     .middleware([configureLogger])
+    .command(health)
     .command(start)
     .command(migrate)
     .fail(handleError)